task2 + переименование старых пакетов
