+++ marks.txt
@@ -1,0 +1,1 @@
+как в junit задать конкретно ожидаемый результат на тестовые данные из MethodSource
