--- src/main/java/Shmidt/lesson16/Day.java
@@ -1,7 +1,0 @@
-package Shmidt.lesson16;
-
-public class Day {
-    public String dayOfWeek;
-    public int dayNumber;
-}
-//Создать класс Day c полем String dayOfWeek и поле с номером относительно начала месяца (не статические).
