+++ src/main/java/Shmidt/lesson2_16_DaysOfWeek/DaysOfWeek.java
@@ -1,0 +1,12 @@
+package Shmidt.lesson2_16_DaysOfWeek;
+
+public class DaysOfWeek {
+    static final String monday ="понедельник";
+    static final String tuesday ="вторник";
+    static final String wednesday ="среда";
+    static final String thursday ="четверг";
+    static final String friday ="пятница";
+    static final String saturday ="суббота";
+    static final String sunday ="воскресенье";
+}
+//    Создать класс DaysOfWeek cо статическими финальными полями - днями недель (Значения: "Понедельник", "Вторник" и тд)
