+++ src/main/java/Shmidt/lesson2_47_usingCycles/task2/Task2.java
@@ -1,0 +1,37 @@
+package Shmidt.lesson2_47_usingCycles.task2;
+
+public class Task2 {
+    public static void main(String[] args) {
+
+        int n = Integer.parseInt(args[0]),//кол-во сенокосилок
+                m = Integer.parseInt(args[1]);//сколько часов косила первая
+        if (m >= 1 && n >= 1)
+            System.out.println("Количество сенокосилок: " + n + "\n" +
+                    "Сколько часов косила первая сенокосилка: " + m);
+
+        int minutes = m * 60;
+        int totalMinutes = 0;
+        for (int i = 1; i <= n; i++) {
+            totalMinutes += minutes;
+            minutes += 10;
+        }
+        System.out.println("Бригада работала " + totalMinutes / 60 + " часов, " + totalMinutes % 60 + " минут");
+    }
+}
+//120 + 130 + 140 + 150 = 540
+/*
+В бригаде, работающей на уборке сена, имеется n сенокосилок.
+
+Первая сенокосилка работала m _часов_, а каждая следующая
+на 10 _минут_ больше, чем предыдущая.
+
+Напишите программу, вычисляющую количество сколько часов
+проработала вся бригада.
+
+Результат вычислений выведите на экран в часах и минутах.
+
+Например, "Бригада работало 0 часов 30 минут".
+
+Значения n и m программа получает через аргументы командной строки.
+Сделайте проверку, что n и m больше нуля. В противном случае выведите сообщение об ошибке и завершите программу.
+*/
