+++ src/test/java/lesson11_3_TestUI/task3/Authorization.java
@@ -4,6 +4,8 @@ import com.codeborne.selenide.Configuration;
 import lesson11_3_TestUI.task3.LoginPage.LoginPage;
 
 import static com.codeborne.selenide.Selenide.open;
+import static com.codeborne.selenide.Selenide.webdriver;
+import static com.codeborne.selenide.WebDriverConditions.url;
 
 public class Authorization {
     final private static String URL = "https://www.saucedemo.com/";
@@ -15,5 +17,6 @@ public class Authorization {
                 .fillLogin(user.getLogin())
                 .fillPassword(user.getPassword())
                 .loginButtonClick();
+        webdriver().shouldHave(url("https://www.saucedemo.com/inventory.html"));
     }
 }
